staticfiles/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-filter==24.3
django-parler==2.3
django-rosetta==0.10.0
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
polib==1.2.0
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
